Refactor for Simplification
