+++ src/main/java/net/countered/terrainslabs/block/ontopofslabs/FlowerOnTop.java
@@ -1,7 +1,6 @@
 package net.countered.terrainslabs.block.ontopofslabs;
 
 import net.countered.terrainslabs.block.interfaces.IBlockCopyFabric;
-import net.countered.terrainslabs.block.interfaces.IOnTopCopy;
 import net.minecraft.block.*;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Vec3d;
@@ -9,7 +8,7 @@ import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.world.BlockView;
 
 
-public class FlowerOnTop extends FlowerBlock implements IBlockCopyFabric, IOnTopCopy {
+public class FlowerOnTop extends FlowerBlock implements IBlockCopyFabric {
     private final Block originalBlock;
     protected static final VoxelShape SHAPE = Block.createCuboidShape(5.0, -8.0, 5.0, 11.0, 2.0, 11.0);
