+++ src/main/java/net/countered/terrainslabs/block/ontopofslabs/GrassOnTop.java
@@ -1,13 +1,12 @@
 package net.countered.terrainslabs.block.ontopofslabs;
 
 import net.countered.terrainslabs.block.interfaces.IBlockCopyFabric;
-import net.countered.terrainslabs.block.interfaces.IOnTopCopy;
 import net.minecraft.block.*;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.world.BlockView;
 
-public class GrassOnTop extends FernBlock implements IBlockCopyFabric, IOnTopCopy {
+public class GrassOnTop extends FernBlock implements IBlockCopyFabric {
     private final Block originalBlock;
     protected static final VoxelShape SHAPE = Block.createCuboidShape(2.0, -8.0, 2.0, 14.0, 5.0, 14.0);
 
@@ -28,5 +27,10 @@ public class GrassOnTop extends FernBlock implements IBlockCopyFabric, IOnTopCop
     @Override
     public Block getOriginBlock() {
         return originalBlock;
+    }
+
+    @Override
+    public BlockCopyType getCopyType() {
+        return BlockCopyType.ON_TOP;
     }
 }
