+++ src/main/java/net/countered/terrainslabs/block/ontopofslabs/SeagrassOnTop.java
@@ -1,15 +1,12 @@
 package net.countered.terrainslabs.block.ontopofslabs;
 
-
 import net.countered.terrainslabs.block.interfaces.IBlockCopyFabric;
-import net.countered.terrainslabs.block.interfaces.IOnTopCopy;
 import net.minecraft.block.*;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.world.BlockView;
 
-
-public class SeagrassOnTop extends SeagrassBlock implements IBlockCopyFabric, IOnTopCopy {
+public class SeagrassOnTop extends SeagrassBlock implements IBlockCopyFabric {
     protected static final VoxelShape SHAPE = Block.createCuboidShape(2.0, -8.0, 2.0, 14.0, 4.0, 14.0);
     private final Block originalBlock;
 
@@ -29,5 +26,10 @@ public class SeagrassOnTop extends SeagrassBlock implements IBlockCopyFabric, IO
     @Override
     public Block getOriginBlock() {
         return originalBlock;
+    }
+
+    @Override
+    public BlockCopyType getCopyType() {
+        return BlockCopyType.ON_TOP;
     }
 }
